refactor(reporter)!: move channels into channels/ submodule
Creates reporter/channels/ subpackage and relocates the channel
implementations, shared HTTP transport, and git info helper:

  channel_base.py         → channels/base.py
  channel_{github,...}.py → channels/{github,...}.py
  _http.py                → channels/_http.py
  _git_info.py            → channels/_git_info.py

Tests mirror the new layout under tests/unit/test_reporter/test_channels/.

_templates/ stays at reporter/ root — it is formatting.py's resource,
not a channel resource.

This commit is layout-only: no behavior change, send(markdown, config)
API preserved. Phase C will refactor the interface.

BREAKING CHANGE: import paths change.
  from ac_guard.reporter.channel_base       → ...channels.base
  from ac_guard.reporter.channel_<platform> → ...channels.<platform>
  from ac_guard.reporter._http              → ...channels._http
  from ac_guard.reporter._git_info          → ...channels._git_info
