+++ src/ac_guard/reporter/channels/__init__.py
@@ -1,0 +1,1 @@
+"""Reporter channels — physical output destinations for rendered reports."""
