+++ src/ac_guard/reporter/channels/_git_info.py
