+++ src/ac_guard/reporter/channels/gitea.py
@@ -1,19 +1,18 @@
-"""GitHub ReportChannel — posts check reports as PR comments.
+"""Gitea ReportChannel — posts check reports as PR comments.
 
-Uses the GitHub REST API to create issue comments on pull requests.
-Supports both github.com and self-hosted GitHub Enterprise instances
+Uses the Gitea REST API to create issue comments on pull requests.
+Supports both gitea.com and self-hosted Gitea instances
 via the ``api_url`` configuration.
 """
 
 from __future__ import annotations
 
 import os
-import re
 from typing import TYPE_CHECKING
 
-from ac_guard.reporter._git_info import get_current_branch, get_remote_repo
-from ac_guard.reporter._http import get_json, post_json
-from ac_guard.reporter.channel_base import (
+from ac_guard.reporter.channels._git_info import get_current_branch, get_remote_repo
+from ac_guard.reporter.channels._http import get_json, post_json
+from ac_guard.reporter.channels.base import (
     ChannelError,
     NoPrContextError,
     ReportChannel,
@@ -23,37 +22,33 @@ from ac_guard.reporter.channel_base import (
 if TYPE_CHECKING:
     from ac_guard.config.models import PrReportConfig
 
-__all__ = ["GitHubChannel"]
-
-_PR_REF_PATTERN = re.compile(r"^refs/pull/(\d+)/")
-"""Pattern to extract PR number from GITHUB_REF."""
+__all__ = ["GiteaChannel"]
 
 
 @register_channel
-class GitHubChannel(ReportChannel):
-    """Post check reports to GitHub PR comments.
+class GiteaChannel(ReportChannel):
+    """Post check reports to Gitea PR comments.
 
     Repository and PR number are resolved automatically:
 
     **Repository** (in priority order):
-        1. ``GITHUB_REPOSITORY`` environment variable
+        1. ``GITEA_REPOSITORY`` environment variable
         2. ``git remote get-url origin`` parsed
 
     **PR number** (in priority order):
         1. ``AI_GUARD_PR_NUMBER`` environment variable
-        2. ``GITHUB_REF`` (``refs/pull/<n>/merge``)
-        3. GitHub API query by current branch name
+        2. Gitea API query by current branch name
 
     Attributes:
-        DEFAULT_API_URL: Default GitHub API base URL.
+        DEFAULT_API_URL: Default Gitea API base URL.
     """
 
-    DEFAULT_API_URL = "https://api.github.com"
+    DEFAULT_API_URL = "https://gitea.com"
 
     @property
     def name(self) -> str:
         """Platform identifier."""
-        return "github"
+        return "gitea"
 
     def send(self, markdown: str, config: PrReportConfig) -> None:
         """Post markdown as a comment on the associated PR.
@@ -71,7 +66,7 @@ class GitHubChannel(ReportChannel):
         api_url = (config.api_url or self.DEFAULT_API_URL).rstrip("/")
         pr_number = self._get_pr_number(token, repo, api_url)
 
-        url = f"{api_url}/repos/{repo}/issues/{pr_number}/comments"
+        url = f"{api_url}/api/v1/repos/{repo}/issues/{pr_number}/comments"
         self._post_json(url, {"body": markdown}, token)
 
     def _get_pr_number(self, token: str, repo: str, api_url: str) -> str:
@@ -79,13 +74,12 @@ class GitHubChannel(ReportChannel):
 
         Priority:
             1. ``AI_GUARD_PR_NUMBER`` env var
-            2. ``GITHUB_REF`` (``refs/pull/<n>/merge``)
-            3. API query by current branch
+            2. API query by current branch
 
         Args:
-            token: GitHub API token for API query fallback.
+            token: Gitea API token for API query fallback.
             repo: Repository in ``owner/repo`` format.
-            api_url: GitHub API base URL.
+            api_url: Gitea API base URL.
 
         Returns:
             PR number as string.
@@ -98,31 +92,23 @@ class GitHubChannel(ReportChannel):
         if pr_number:
             return pr_number
 
-        # 2. GITHUB_REF
-        github_ref = os.environ.get("GITHUB_REF", "")
-        match = _PR_REF_PATTERN.match(github_ref)
-        if match:
-            return match.group(1)
-
-        # 3. API query by branch
+        # 2. API query by branch
         branch = get_current_branch()
         if branch:
-            owner = repo.split("/")[0] if "/" in repo else repo
-            query_url = (
-                f"{api_url}/repos/{repo}/pulls"
-                f"?head={owner}:{branch}&state=open&per_page=1"
-            )
+            query_url = f"{api_url}/api/v1/repos/{repo}/pulls?state=open&limit=50"
             try:
                 data = self._get_json(query_url, token)
-                if data and isinstance(data, list) and len(data) > 0:
-                    return str(data[0]["number"])
+                if data and isinstance(data, list):
+                    for pr in data:
+                        head = pr.get("head", {})
+                        if head.get("label") == branch:
+                            return str(pr["number"])
             except ChannelError:
                 pass  # Fall through to error
 
         raise NoPrContextError(
             "Cannot determine PR number. Set AI_GUARD_PR_NUMBER, "
-            "GITHUB_REF (refs/pull/<n>/merge), or push your branch "
-            "and open a PR"
+            "or push your branch and open a PR"
         )
 
     @staticmethod
@@ -141,7 +127,7 @@ class GitHubChannel(ReportChannel):
         token = os.environ.get(config.token_env)
         if not token:
             raise ChannelError(
-                f"GitHub token not found: set the {config.token_env} "
+                f"Gitea token not found: set the {config.token_env} "
                 f"environment variable"
             )
         return token
@@ -151,7 +137,7 @@ class GitHubChannel(ReportChannel):
         """Determine the repository.
 
         Priority:
-            1. ``GITHUB_REPOSITORY`` env var
+            1. ``GITEA_REPOSITORY`` env var
             2. ``git remote get-url origin``
 
         Returns:
@@ -160,7 +146,7 @@ class GitHubChannel(ReportChannel):
         Raises:
             ChannelError: If repository cannot be determined.
         """
-        repo = os.environ.get("GITHUB_REPOSITORY")
+        repo = os.environ.get("GITEA_REPOSITORY")
         if repo:
             return repo
 
@@ -169,32 +155,28 @@ class GitHubChannel(ReportChannel):
             return repo
 
         raise ChannelError(
-            "Cannot determine repository. Set GITHUB_REPOSITORY or "
+            "Cannot determine repository. Set GITEA_REPOSITORY or "
             "ensure git remote 'origin' is configured"
         )
 
     @staticmethod
     def _post_json(url: str, body: dict, token: str) -> None:
-        """POST JSON with Bearer auth via the shared retry layer."""
+        """POST JSON with token auth via the shared retry layer."""
         post_json(
             url,
             headers={
-                "Authorization": f"Bearer {token}",
-                "Accept": "application/vnd.github+json",
+                "Authorization": f"token {token}",
                 "Content-Type": "application/json",
             },
             body=body,
-            api_name="GitHub",
+            api_name="Gitea",
         )
 
     @staticmethod
     def _get_json(url: str, token: str) -> list | dict | None:
-        """GET JSON with Bearer auth via the shared retry layer."""
+        """GET JSON with token auth via the shared retry layer."""
         return get_json(
             url,
-            headers={
-                "Authorization": f"Bearer {token}",
-                "Accept": "application/vnd.github+json",
-            },
-            api_name="GitHub",
+            headers={"Authorization": f"token {token}"},
+            api_name="Gitea",
         )
