+++ tests/integration/test_m5_report_and_output.py
@@ -19,7 +19,7 @@ from typer.testing import CliRunner
 from ac_guard.checker.core import run_stage
 from ac_guard.cli.main import app
 from ac_guard.config.models import PrReportConfig
-from ac_guard.reporter.channel_base import (
+from ac_guard.reporter.channels.base import (
     ChannelError,
     NoPrContextError,
     post_pr_comment,
@@ -313,7 +313,7 @@ class TestCliPrReportIntegration:
         config = _write_config_pr_report(tmp_path, enabled=True)
         with (
             patch("ac_guard.checker.core.get_changed_files", return_value=[]),
-            patch("ac_guard.reporter.channel_github.GitHubChannel.send") as mock_send,
+            patch("ac_guard.reporter.channels.github.GitHubChannel.send") as mock_send,
         ):
             result = runner.invoke(app, ["check", "--config", str(config)])
         assert result.exit_code == 0
@@ -329,7 +329,7 @@ class TestCliPrReportIntegration:
         config = _write_config_pr_report(tmp_path, enabled=True)
         with (
             patch("ac_guard.checker.core.get_changed_files", return_value=[]),
-            patch("ac_guard.reporter.channel_github.GitHubChannel.send") as mock_send,
+            patch("ac_guard.reporter.channels.github.GitHubChannel.send") as mock_send,
         ):
             result = runner.invoke(
                 app, ["gate", "run", "--stage", "pre-commit", "--config", str(config)]
@@ -343,7 +343,7 @@ class TestCliPrReportIntegration:
         with (
             patch("ac_guard.checker.core.get_changed_files", return_value=[]),
             patch(
-                "ac_guard.reporter.channel_github.GitHubChannel.send",
+                "ac_guard.reporter.channels.github.GitHubChannel.send",
                 side_effect=NoPrContextError("Cannot determine PR number"),
             ),
         ):
@@ -365,7 +365,7 @@ class TestCliPrReportIntegration:
         with (
             patch("ac_guard.checker.core.get_changed_files", return_value=[]),
             patch(
-                "ac_guard.reporter.channel_github.GitHubChannel.send",
+                "ac_guard.reporter.channels.github.GitHubChannel.send",
                 side_effect=ChannelError("GitHub API returned 500"),
             ),
         ):
@@ -387,7 +387,7 @@ class TestCliPrReportIntegration:
         config = _write_config_pr_report(tmp_path, enabled=False)
         with (
             patch("ac_guard.checker.core.get_changed_files", return_value=[]),
-            patch("ac_guard.reporter.channel_github.GitHubChannel.send") as mock_send,
+            patch("ac_guard.reporter.channels.github.GitHubChannel.send") as mock_send,
         ):
             result = runner.invoke(app, ["check", "--config", str(config)])
         assert result.exit_code == 0
