+++ tests/unit/test_reporter/test_channels/test_gitea.py
@@ -8,8 +8,8 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from ac_guard.config.models import PrReportConfig
-from ac_guard.reporter.channel_base import ChannelError, NoPrContextError
-from ac_guard.reporter.channel_gitea import GiteaChannel
+from ac_guard.reporter.channels.base import ChannelError, NoPrContextError
+from ac_guard.reporter.channels.gitea import GiteaChannel
 
 
 class TestGiteaChannel:
@@ -93,7 +93,7 @@ class TestGiteaChannel:
 
         with (
             patch(
-                "ac_guard.reporter.channel_gitea.get_remote_repo",
+                "ac_guard.reporter.channels.gitea.get_remote_repo",
                 return_value="org/my-repo",
             ),
             patch("urllib.request.urlopen", return_value=self._mock_urlopen()) as m,
@@ -126,7 +126,7 @@ class TestGiteaChannel:
 
         with (
             patch(
-                "ac_guard.reporter.channel_gitea.get_current_branch",
+                "ac_guard.reporter.channels.gitea.get_current_branch",
                 return_value="feat/test",
             ),
             patch("urllib.request.urlopen", side_effect=urlopen_side_effect) as m,
@@ -144,7 +144,7 @@ class TestGiteaChannel:
 
         with (
             patch(
-                "ac_guard.reporter.channel_gitea.get_current_branch", return_value=None
+                "ac_guard.reporter.channels.gitea.get_current_branch", return_value=None
             ),
             pytest.raises(ChannelError, match="PR number"),
         ):
@@ -160,7 +160,7 @@ class TestGiteaChannel:
 
         with (
             patch(
-                "ac_guard.reporter.channel_gitea.get_current_branch", return_value=None
+                "ac_guard.reporter.channels.gitea.get_current_branch", return_value=None
             ),
             pytest.raises(NoPrContextError, match="PR number"),
         ):
@@ -179,7 +179,7 @@ class TestGiteaChannel:
         side_effect = [URLError("transient"), self._mock_urlopen()]
         with (
             patch("urllib.request.urlopen", side_effect=side_effect) as m,
-            patch("ac_guard.reporter._http.time.sleep"),
+            patch("ac_guard.reporter.channels._http.time.sleep"),
         ):
             GiteaChannel().send("## Report", self._make_config())
         assert m.call_count == 2
