+++ tests/unit/test_reporter/test_channels/test_gitlab.py
@@ -8,8 +8,8 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from ac_guard.config.models import PrReportConfig
-from ac_guard.reporter.channel_base import ChannelError, NoPrContextError
-from ac_guard.reporter.channel_gitlab import GitLabChannel
+from ac_guard.reporter.channels.base import ChannelError, NoPrContextError
+from ac_guard.reporter.channels.gitlab import GitLabChannel
 
 
 class TestGitLabChannel:
@@ -95,7 +95,7 @@ class TestGitLabChannel:
 
         with (
             patch(
-                "ac_guard.reporter.channel_gitlab.get_remote_repo",
+                "ac_guard.reporter.channels.gitlab.get_remote_repo",
                 return_value="org/my-repo",
             ),
             patch("urllib.request.urlopen", return_value=self._mock_urlopen()) as m,
@@ -124,7 +124,7 @@ class TestGitLabChannel:
 
         with (
             patch(
-                "ac_guard.reporter.channel_gitlab.get_current_branch",
+                "ac_guard.reporter.channels.gitlab.get_current_branch",
                 return_value="feat/test",
             ),
             patch("urllib.request.urlopen", side_effect=urlopen_side_effect) as m,
@@ -143,7 +143,8 @@ class TestGitLabChannel:
 
         with (
             patch(
-                "ac_guard.reporter.channel_gitlab.get_current_branch", return_value=None
+                "ac_guard.reporter.channels.gitlab.get_current_branch",
+                return_value=None,
             ),
             pytest.raises(ChannelError, match="MR IID"),
         ):
@@ -160,7 +161,8 @@ class TestGitLabChannel:
 
         with (
             patch(
-                "ac_guard.reporter.channel_gitlab.get_current_branch", return_value=None
+                "ac_guard.reporter.channels.gitlab.get_current_branch",
+                return_value=None,
             ),
             pytest.raises(NoPrContextError, match="MR IID"),
         ):
@@ -179,7 +181,7 @@ class TestGitLabChannel:
         side_effect = [URLError("transient"), self._mock_urlopen()]
         with (
             patch("urllib.request.urlopen", side_effect=side_effect) as m,
-            patch("ac_guard.reporter._http.time.sleep"),
+            patch("ac_guard.reporter.channels._http.time.sleep"),
         ):
             GitLabChannel().send("## Report", self._make_config())
         assert m.call_count == 2
